requirements.txt
Django
Pillow
Fabric
argparse
boto
coverage
django-debug-toolbar
greeking
flake8
psycopg
# pycrypto (Unmantained)
pycryptodome
pyflakes
python-coveralls
python-memcached
pytz
six
sqlparse
django-markdownify
facebook-sdk
django-environ
social-auth-app-django
facepy
django-crontab
beautifulsoup4
markdown
requests
requests-cache
django-request
django-smuggler
django-defender
tarsafe
django-modeltranslation
